gets all resources on initial page #87
